fix(convex): isolate remote assist query reads
