+++ packages/athena-webapp/convex/remoteAssist/public.ts
@@ -23,6 +23,7 @@ import {
   startRemoteAssistSession,
 } from "./application/sessionService";
 import type { RemoteAssistSession } from "./application/types";
+import { createRemoteAssistReadRepository } from "./infrastructure/remoteAssistReadRepository";
 import { createRemoteAssistRepository } from "./infrastructure/remoteAssistRepository";
 
 const remoteAssistClientReturnValidator = v.object({
@@ -92,7 +93,7 @@ export const getClientByRuntime = query({
       userId: athenaUser._id,
     });
 
-    const repository = createRemoteAssistRepository(ctx);
+    const repository = createRemoteAssistReadRepository(ctx);
     return repository.getClientByRuntime(args);
   },
 });
@@ -104,7 +105,7 @@ export const getCurrentSessionByClient = query({
   returns: v.union(remoteAssistSessionReturnValidator, v.null()),
   handler: async (ctx, args) => {
     const athenaUser = await requireAuthenticatedAthenaUserWithCtx(ctx);
-    const repository = createRemoteAssistRepository(ctx);
+    const repository = createRemoteAssistReadRepository(ctx);
     const client = await repository.getClient(args.clientId);
     if (!client) {
       return null;
